feat: 수정 요청 사항 반영을 위한 비즈니스 로직 수정 및 Service, Impl 분리 (#241)
