+++ server/module-core/src/main/java/com/momo/domain/schedule/service/ScheduleService.java
@@ -1,12 +1,5 @@
 package com.momo.domain.schedule.service;
 
-import com.momo.domain.common.exception.CustomException;
-import com.momo.domain.common.exception.ErrorCode;
-import com.momo.domain.group.entity.Group;
-import com.momo.domain.group.repository.GroupRepository;
-import com.momo.domain.group.repository.ParticipantRepository;
-import com.momo.domain.schedule.entity.Schedule;
-import com.momo.domain.schedule.repository.ScheduleRepository;
 import com.momo.domain.schedule.dto.GroupScheduleResponse;
 import com.momo.domain.schedule.dto.GroupScheduleResponses;
 import com.momo.domain.schedule.dto.GroupSchedulesRequest;
@@ -14,64 +7,13 @@ import com.momo.domain.schedule.dto.ScheduleCreateRequest;
 import com.momo.domain.schedule.dto.UserScheduleResponse;
 import com.momo.domain.schedule.dto.UserSchedulesRequest;
 import com.momo.domain.user.entity.User;
-import java.time.LocalDateTime;
 import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class ScheduleService {
+public interface ScheduleService {
 
-    private final ScheduleRepository scheduleRepository;
+    GroupScheduleResponse create(User user, ScheduleCreateRequest request);
 
-    private final GroupRepository groupRepository;
+    GroupScheduleResponses findPageByUserAndGroupId(User user, GroupSchedulesRequest request);
 
-    private final ParticipantRepository participantRepository;
-
-    public Long create(User user, ScheduleCreateRequest request) {
-        Group group = getGroupById(request.getGroupId());
-        validateGroupManager(group, user);
-        Schedule schedule = Schedule.create(request.toEntity(), group, user);
-        return scheduleRepository.save(schedule).getId();
-    }
-
-    public void validateGroupManager(Group group, User user) {
-        if (!group.isManager(user)) {
-            throw new CustomException(ErrorCode.GROUP_MANAGER_AUTHORIZED);
-        }
-    }
-
-    @Transactional(readOnly = true)
-    public GroupScheduleResponses findPageByUserAndGroupId(User user, GroupSchedulesRequest request) {
-        Group group = getGroupById(request.getGroupId());
-        validateParticipant(group, user);
-        PageRequest pageRequest = PageRequest.of(request.getPage(), request.getSize());
-        List<GroupScheduleResponse> responses = scheduleRepository
-            .findAllByGroupAndUserOrderByCreatedDateDesc(group, user.getId(), pageRequest);
-        return GroupScheduleResponses.of(responses, group.getManager().getId());
-    }
-
-    public void validateParticipant(Group group, User user) {
-        if (!participantRepository.existsByGroupAndUser(group, user)) {
-            throw new CustomException(ErrorCode.GROUP_PARTICIPANT_UNAUTHORIZED);
-        }
-    }
-
-    @Transactional(readOnly = true)
-    public List<UserScheduleResponse> findPageByUserAndSearchDate(User user, UserSchedulesRequest request) {
-        LocalDateTime searchStartDateTime = request.getSearchStartDate().atStartOfDay();
-        LocalDateTime searchEndDateTime = request.getSearchEndDate().plusDays(1).atStartOfDay();
-        List<Schedule> schedules = scheduleRepository
-            .findAllByStartDateTimeBetween(searchStartDateTime, searchEndDateTime, user);
-        return UserScheduleResponse.listOf(schedules);
-    }
-
-    public Group getGroupById(Long groupId) {
-        return groupRepository.findById(groupId)
-            .orElseThrow(() -> new CustomException(ErrorCode.INVALID_INDEX_NUMBER));
-    }
+    List<UserScheduleResponse> findPageByUserAndSearchDate(User user, UserSchedulesRequest request);
 }
